parsers: split ParserBase into ParserBase and LengthConstrainedParserBase
The former has a name field only, the latter has the min_len and max_len
fields.
